教师用户的请假申请，机构管理员对请假的审批

+++ BearKidDiaryWeb/src/com/bearkiddiary/dao/UserDao.java
@@ -19,7 +19,7 @@ public interface UserDao extends BaseDao<User> {
 	 * @param user
 	 * @return
 	 */
-	int Register(User user);
+	Long Register(User user);
 
 	/**
 	 * 检测该号码是否被注册过
